Moved core.template and core.websocket to tornadoweb module.
Refs: #109
